vmm: uintptr_t for the physical-frame allocator API (cross-arch Phase 7, 1/n)
Retype the shared frame-allocator surface that the future aarch64 VMM also
calls: vmm_find_free_page (returns a phys addr), free_page, adjust_cow_counter,
and vmm_share_ref now take/return uintptr_t instead of u_int32_t.  On i386
uintptr_t == unsigned int, so this is codegen-neutral (verified: builds + boots
unchanged); on a 64-bit target these stop truncating physical addresses.
Internal frame indices stay u_int32_t (they are counts, not addresses).

+++ sys/include/vmm/vmm.h
@@ -71,10 +71,10 @@ extern "C" {
     int vmm_init();
     int vmm_mem_map_init();
     u_int32_t count_memory();
-    u_int32_t vmm_find_free_page(pidType pid);
-    int free_page(u_int32_t pageAddr);
-    int adjust_cow_counter(u_int32_t baseAddr, int adjustment);
-    void vmm_share_ref(u_int32_t phys);
+    uintptr_t vmm_find_free_page(pidType pid);
+    int free_page(uintptr_t pageAddr);
+    int adjust_cow_counter(uintptr_t baseAddr, int adjustment);
+    void vmm_share_ref(uintptr_t phys);
     void vmm_free_process_pages(pidType pid);
     u_int32_t vmm_audit_orphan_pages(void);
